+++ engine/apps/api/serializers/on_call_shifts.py
@@ -1,6 +1,6 @@
 from rest_framework import serializers
 
-from apps.schedules.models import CustomOnCallShift, OnCallScheduleWeb
+from apps.schedules.models import CustomOnCallShift, OnCallSchedule
 from apps.user_management.models import User
 from common.api_helpers.custom_fields import (
     OrganizationFilteredPrimaryKeyRelatedField,
@@ -18,7 +18,7 @@ class OnCallShiftSerializer(EagerLoadingMixin, serializers.ModelSerializer):
         required=True,
         choices=CustomOnCallShift.WEB_TYPES,
     )
-    schedule = OrganizationFilteredPrimaryKeyRelatedField(queryset=OnCallScheduleWeb.objects)
+    schedule = OrganizationFilteredPrimaryKeyRelatedField(queryset=OnCallSchedule.objects)
     frequency = serializers.ChoiceField(required=False, choices=CustomOnCallShift.FREQUENCY_CHOICES, allow_null=True)
     shift_start = serializers.DateTimeField(source="start")
     shift_end = serializers.SerializerMethodField()
@@ -174,7 +174,7 @@ class OnCallShiftSerializer(EagerLoadingMixin, serializers.ModelSerializer):
         )
         instance = super().create(validated_data)
 
-        instance.start_drop_ical_and_check_schedule_tasks(instance.schedule)
+        instance.start_drop_ical_and_check_schedule_tasks(instance.schedule, source=instance.source)
         return instance
 
 
@@ -208,5 +208,5 @@ class OnCallShiftUpdateSerializer(OnCallShiftSerializer):
         else:
             result = super().update(instance, validated_data)
 
-        instance.start_drop_ical_and_check_schedule_tasks(instance.schedule)
+        instance.start_drop_ical_and_check_schedule_tasks(instance.schedule, source=instance.source)
         return result