+++ grafana-plugin/src/pages/schedule/Schedule.tsx
@@ -108,10 +108,14 @@ class SchedulePage extends React.Component<SchedulePageProps, SchedulePageState>
     const users = store.userStore.getSearchResult().results;
     const schedule = scheduleStore.items[scheduleId];
 
-    const disabled =
+    const disabledRotationForm =
       !isUserActionAllowed(UserActions.SchedulesWrite) ||
       schedule?.type !== ScheduleType.API ||
-      shiftIdToShowRotationForm ||
+      shiftIdToShowRotationForm;
+
+    const disabledOverrideForm =
+      !isUserActionAllowed(UserActions.SchedulesWrite) ||
+      (schedule?.type !== ScheduleType.API && schedule?.type !== ScheduleType.Calendar) ||
       shiftIdToShowOverridesForm;
 
     return (
@@ -214,7 +218,7 @@ class SchedulePage extends React.Component<SchedulePageProps, SchedulePageState>
                     currentTimezone={currentTimezone}
                     startMoment={startMoment}
                     onClick={this.handleShowForm}
-                    disabled={disabled}
+                    disabled={disabledRotationForm}
                   />
                   <Rotations
                     scheduleId={scheduleId}
@@ -225,7 +229,7 @@ class SchedulePage extends React.Component<SchedulePageProps, SchedulePageState>
                     onDelete={this.handleDeleteRotation}
                     shiftIdToShowRotationForm={shiftIdToShowRotationForm}
                     onShowRotationForm={this.handleShowRotationForm}
-                    disabled={disabled}
+                    disabled={disabledRotationForm}
                   />
                   <ScheduleOverrides
                     scheduleId={scheduleId}
@@ -236,7 +240,7 @@ class SchedulePage extends React.Component<SchedulePageProps, SchedulePageState>
                     onDelete={this.handleDeleteOverride}
                     shiftIdToShowRotationForm={shiftIdToShowOverridesForm}
                     onShowRotationForm={this.handleShowOverridesForm}
-                    disabled={disabled}
+                    disabled={disabledOverrideForm}
                   />
                 </div>
               </VerticalGroup>